[fix] Move ViewerToolbar to components folder
